Fix view-range pop, KidsMode shadow pulse, clang-tidy pass; rollback LOD terrain
- renderd3d: revert ProcessMapMerged LOD tile-merge attempt back to pre-LOD
  ring loop; document limitations in ROADMAP.md
- renderd3d: smooth object and creature fade at view-range edge using
  Chebyshev tile distance with quadratic falloff instead of hard pop
- renderd3d: fix KidsMode dead-dino shadow pulsing with breathing animation
  by guarding the (at-FTime)/at fade behind !KidsMode
- mathematics: remove UB NormVector(l, 1.0f) call on uninitialised variable
- Game: fix self-init UB (TimeDt = TimeDt -> ::TimeDt); atoi/atof -> strtol/strtod
- Resources: replace all atoi/atof calls with strtol/strtod
- DataDefs: size JKind enum base to uint8_t
- clang-tidy: auto-fix 436 float-promotion warnings (fabs->fabsf, sin->sinf etc.)
- clang-tidy: NOLINT annotations for int-to-ptr, static-init globals, sscanf
- ShaderPack: shader pack loader with VFS integration; ship built-in packs
  (cinematic, god-rays, heightfog, shadows, ssao, water)

+++ renderer/PostProcessing.cpp
@@ -196,7 +196,7 @@ bool PostProcessingPipeline::Initialize(int screenWidth, int screenHeight) {
     glBindBuffer(GL_ARRAY_BUFFER, m_fsQuadVbo);
     glBufferData(GL_ARRAY_BUFFER, sizeof(quadVertices), quadVertices, GL_STATIC_DRAW);
     glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE, 4 * sizeof(float), nullptr);
-    glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 4 * sizeof(float), (void*)(2 * sizeof(float)));
+    glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 4 * sizeof(float), (void*)(uintptr_t)(2 * sizeof(float))); // NOLINT(performance-no-int-to-ptr)
     glEnableVertexAttribArray(0);
     glEnableVertexAttribArray(1);
     glBindBuffer(GL_ARRAY_BUFFER, 0);
